Ran R# code cleanup

+++ Sources/BandcampDownloader/Windows/MainWindow.xaml.cs
@@ -1,19 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Media;
-using System.Net;
-using System.Reflection;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Documents;
+using System.Windows.Forms;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Shell;
@@ -21,12 +13,13 @@ using BandcampDownloader.MediatorMessages;
 using BandcampDownloader.ViewModels;
 using Config.Net;
 using DevExpress.Mvvm;
-using ImageResizer;
+using Cursors = System.Windows.Input.Cursors;
+using MessageBox = System.Windows.MessageBox;
 
-namespace BandcampDownloader {
-
-    public partial class MainWindow: Window {
-
+namespace BandcampDownloader
+{
+    public partial class MainWindow : Window
+    {
         //TODO - Remove the view model hacks, these were just done for experimental purposes!
 
 
@@ -52,12 +45,13 @@ namespace BandcampDownloader {
         }
 
         /// <summary>
-        /// Updates the state of the controls.
+        ///     Updates the state of the controls.
         /// </summary>
         /// <param name="downloadStarted">True if the download just started, false if it just stopped.</param>
         private void UpdateControlsState(bool downloadStarted)
         {
-            this.Dispatcher.Invoke(new Action(() => {
+            Dispatcher.Invoke(() =>
+            {
                 if (downloadStarted)
                 {
                     // We just started the download
@@ -89,7 +83,8 @@ namespace BandcampDownloader {
                     buttonStop.IsEnabled = false;
                     buttonBrowse.IsEnabled = true;
                     textBoxUrls.IsReadOnly = false;
-                    progressBar.Foreground = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF01D328")); // Green
+                    progressBar.Foreground =
+                        new SolidColorBrush((Color) ColorConverter.ConvertFromString("#FF01D328")); // Green
                     progressBar.IsIndeterminate = false;
                     progressBar.Value = progressBar.Minimum;
                     TaskbarItemInfo.ProgressState = TaskbarItemProgressState.None;
@@ -106,23 +101,21 @@ namespace BandcampDownloader {
                     textBoxCoverArtMaxSize.IsEnabled = true;
                     checkBoxRetrieveFilesizes.IsEnabled = true;
                 }
-            }));
+            });
         }
 
 
         /// <summary>
-        /// Displays the specified message in the log with the specified color.
+        ///     Displays the specified message in the log with the specified color.
         /// </summary>
         /// <param name="message">The message.</param>
         /// <param name="color">The color.</param>
         private void Log(DownloaderLogMessage message)
         {
-            Dispatcher.Invoke(() => {
-
-                if (!((MainViewModel)DataContext).UserSettings.ShowVerboseLog && (message.LogType == LogType.Warning || message.LogType == LogType.VerboseInfo))
-                {
-                    return;
-                }
+            Dispatcher.Invoke(() =>
+            {
+                if (!((MainViewModel) DataContext).UserSettings.ShowVerboseLog &&
+                    (message.LogType == LogType.Warning || message.LogType == LogType.VerboseInfo)) return;
 
                 // Time
                 var textRange = new TextRange(richTextBoxLog.Document.ContentEnd, richTextBoxLog.Document.ContentEnd);
@@ -135,10 +128,7 @@ namespace BandcampDownloader {
                 // Line break
 
                 richTextBoxLog.AppendText(Environment.NewLine);
-                if (((MainViewModel)DataContext).UserSettings.AutoScrollLog)
-                {
-                    richTextBoxLog.ScrollToEnd();
-                }
+                if (((MainViewModel) DataContext).UserSettings.AutoScrollLog) richTextBoxLog.ScrollToEnd();
             });
         }
 
@@ -149,8 +139,7 @@ namespace BandcampDownloader {
 
         private void PlayASound()
         {
-
-            (new SoundPlayer(@"C:\Windows\Media\Windows Ding.wav")).Play();
+            new SoundPlayer(@"C:\Windows\Media\Windows Ding.wav").Play();
         }
 
         private void ProcessProgressBarMessage(ProgressUpdateMessage message)
@@ -163,38 +152,38 @@ namespace BandcampDownloader {
         #endregion
 
 
-
         #region Events
 
-        private void buttonBrowse_Click(object sender, RoutedEventArgs e) {
-            var dialog = new System.Windows.Forms.FolderBrowserDialog();
+        private void buttonBrowse_Click(object sender, RoutedEventArgs e)
+        {
+            var dialog = new FolderBrowserDialog();
             dialog.Description = "Select the folder to save albums";
-            if (dialog.ShowDialog() == System.Windows.Forms.DialogResult.OK) {
+            if (dialog.ShowDialog() == System.Windows.Forms.DialogResult.OK)
                 textBoxDownloadsLocation.Text = dialog.SelectedPath;
-            }
         }
 
-        private void buttonDefaultSettings_Click(object sender, RoutedEventArgs e) {
-            if (MessageBox.Show("Reset settings to their default values?", "Bandcamp Downloader", MessageBoxButton.OKCancel, MessageBoxImage.Question, MessageBoxResult.Cancel) == MessageBoxResult.OK) {
-                InitializeSettings(true);
-            }
+        private void buttonDefaultSettings_Click(object sender, RoutedEventArgs e)
+        {
+            if (MessageBox.Show("Reset settings to their default values?", "Bandcamp Downloader",
+                    MessageBoxButton.OKCancel, MessageBoxImage.Question, MessageBoxResult.Cancel) ==
+                MessageBoxResult.OK) InitializeSettings(true);
         }
 
-        private void buttonStop_Click(object sender, RoutedEventArgs e) {
-            if (MessageBox.Show("Would you like to cancel all downloads?", "Bandcamp Downloader", MessageBoxButton.YesNo, MessageBoxImage.Question, MessageBoxResult.No) != MessageBoxResult.Yes) {
-                return;
-            }
+        private void buttonStop_Click(object sender, RoutedEventArgs e)
+        {
+            if (MessageBox.Show("Would you like to cancel all downloads?", "Bandcamp Downloader",
+                    MessageBoxButton.YesNo, MessageBoxImage.Question, MessageBoxResult.No) !=
+                MessageBoxResult.Yes) return;
 
             Dispatcher.Invoke(() =>
             {
-
-                ((MainViewModel)DataContext).UserCanceled = true;
+                ((MainViewModel) DataContext).UserCanceled = true;
                 Cursor = Cursors.Wait;
                 Log(new DownloaderLogMessage("Cancelling downloads. Please wait...", LogType.Info));
 
-                lock (((MainViewModel)DataContext).PendingDownloads)
+                lock (((MainViewModel) DataContext).PendingDownloads)
                 {
-                    if (((MainViewModel)DataContext).PendingDownloads.Count == 0)
+                    if (((MainViewModel) DataContext).PendingDownloads.Count == 0)
                     {
                         // Nothing to cancel
                         Cursor = Cursors.Arrow;
@@ -203,52 +192,51 @@ namespace BandcampDownloader {
                 }
 
                 buttonStop.IsEnabled = false;
-                progressBar.Foreground = System.Windows.Media.Brushes.Red;
+                progressBar.Foreground = Brushes.Red;
                 progressBar.IsIndeterminate = true;
                 TaskbarItemInfo.ProgressState = TaskbarItemProgressState.None;
                 TaskbarItemInfo.ProgressValue = 0;
 
-                lock (((MainViewModel)DataContext).PendingDownloads)
+                lock (((MainViewModel) DataContext).PendingDownloads)
                 {
                     // Stop current downloads
-                    foreach (WebClient webClient in ((MainViewModel)DataContext).PendingDownloads)
-                    {
-                        webClient.CancelAsync();
-                    }
+                    foreach (var webClient in ((MainViewModel) DataContext).PendingDownloads) webClient.CancelAsync();
                 }
 
                 Cursor = Cursors.Arrow;
-
             });
-
-
         }
 
-        private void InitializeSettings(Boolean resetToDefaults) {
-            if (resetToDefaults) {
-                File.Delete(Constants.UserSettingsFilePath);
-            }
+        private void InitializeSettings(bool resetToDefaults)
+        {
+            if (resetToDefaults) File.Delete(Constants.UserSettingsFilePath);
             // Must set this before UI forms, cannot default in settings as it isn't determined by a constant function
-            if (String.IsNullOrEmpty(((MainViewModel)DataContext).UserSettings.DownloadsLocation)) {
-                ((MainViewModel)DataContext).UserSettings.DownloadsLocation = Environment.GetFolderPath(Environment.SpecialFolder.Desktop) + "\\{artist}\\{album}";
-            }
-            ((MainViewModel)DataContext).UserSettings = new ConfigurationBuilder<UserSettings>().UseIniFile(Constants.UserSettingsFilePath).Build();
+            if (string.IsNullOrEmpty(((MainViewModel) DataContext).UserSettings.DownloadsLocation))
+                ((MainViewModel) DataContext).UserSettings.DownloadsLocation =
+                    Environment.GetFolderPath(Environment.SpecialFolder.Desktop) + "\\{artist}\\{album}";
+            ((MainViewModel) DataContext).UserSettings = new ConfigurationBuilder<UserSettings>()
+                .UseIniFile(Constants.UserSettingsFilePath).Build();
         }
 
-        private void labelVersion_MouseDown(object sender, MouseButtonEventArgs e) {
+        private void labelVersion_MouseDown(object sender, MouseButtonEventArgs e)
+        {
             Process.Start(Constants.ProjectWebsite);
         }
 
-        private void textBoxUrls_GotFocus(object sender, RoutedEventArgs e) {
-            if (textBoxUrls.Text == Constants.UrlsHint) {
+        private void textBoxUrls_GotFocus(object sender, RoutedEventArgs e)
+        {
+            if (textBoxUrls.Text == Constants.UrlsHint)
+            {
                 // Erase the hint message
                 textBoxUrls.Text = "";
                 textBoxUrls.Foreground = new SolidColorBrush(Colors.Black);
             }
         }
 
-        private void textBoxUrls_LostFocus(object sender, RoutedEventArgs e) {
-            if (textBoxUrls.Text == "") {
+        private void textBoxUrls_LostFocus(object sender, RoutedEventArgs e)
+        {
+            if (textBoxUrls.Text == "")
+            {
                 // Show the hint message
                 textBoxUrls.Text = Constants.UrlsHint;
                 textBoxUrls.Foreground = new SolidColorBrush(Colors.DarkGray);
@@ -257,20 +245,15 @@ namespace BandcampDownloader {
 
         private void WindowMain_Closing(object sender, CancelEventArgs e)
         {
-
             Dispatcher.Invoke(() =>
             {
-                if (((MainViewModel)DataContext).ActiveDownloads)
-                {
-                    // There are active downloads, ask for confirmation
-                    if (MessageBox.Show("There are currently active downloads. Are you sure you want to close the application and stop all downloads?", "Bandcamp Downloader", MessageBoxButton.OKCancel, MessageBoxImage.Warning, MessageBoxResult.Cancel) == MessageBoxResult.Cancel)
-                    {
-                        // Cancel closing the window
+                if (((MainViewModel) DataContext).ActiveDownloads)
+                    if (MessageBox.Show(
+                            "There are currently active downloads. Are you sure you want to close the application and stop all downloads?",
+                            "Bandcamp Downloader", MessageBoxButton.OKCancel, MessageBoxImage.Warning,
+                            MessageBoxResult.Cancel) == MessageBoxResult.Cancel)
                         e.Cancel = true;
-                    }
-                }
             });
-
         }
 
         #endregion Events
